Write a program that multiplies two numbers.
    Your program should print the result of the multiplication, followed by a new line
    You can assume that the two numbers and result of the multiplication can be stored in an integer
    If the program does not receive two arguments, your program should print Error, followed by a new line, and return 1

+++ argc_argv/3-mul.c
@@ -13,8 +13,8 @@
 
 int main(int argc, char **argv)
 {
-	int a = 0;
-	int b = 0;
+	long a = 0;
+	long b = 0;
 	int i = 0;
 	int j = 0;
 	
@@ -36,7 +36,7 @@ int main(int argc, char **argv)
 			j++;
 		}
 
-		printf("%i\n", a * b);
+		printf("%ld\n", a * b);
 	}
 
 	return (0);
